Fixed missing namespaces
